backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
diffusers==0.24.0
transformers==4.35.2
accelerate==0.24.1
xformers==0.0.22
safetensors==0.4.0 